src/requirements.txt
fastapi
uvicorn[standard]
python-gitlab
openai
